+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [2021-01-10] Allow overriding of reported address to backend by [@lflare].
 
 ### Changed
 - [2021-01-10] Increased interval of refresh & backend ping to 30 seconds by [@lflare].